requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
albucore==0.0.12
albumentations==1.4.11
alembic==1.13.0
annotated-types==0.7.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
catalyst==20.9
causal-conv1d==1.2.0.post2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
databricks-cli==0.18.0
datasets==2.15.0
decorator==5.1.1
deprecation==2.1.0
dill==0.3.7
docker==6.1.3
einops==0.7.0
entrypoints==0.4
eval_type_backport==0.2.0
executing==2.0.1
filelock==3.13.1
Flask==3.0.0
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.10.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.2
grpcio==1.65.1
gunicorn==21.2.0
huggingface-hub==0.19.4
idna==3.6
imageio==2.34.2
importlib-metadata==7.0.0
iopath==0.1.10
ipython==8.26.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.4
lightning==2.3.3
lightning-utilities==0.11.5
Mako==1.3.0
mamba-ssm @ file:///root/data/model/urbanssf-new/mamba_ssm-1.2.0.post1%2Bcu122torch2.1cxx11abiFALSE-cp311-cp311-linux_x86_64.whl#sha256=edba5bea9e9afdcb173f7bd52809cc25e6134f80d618f1ec3619374d89ee9493
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.7
mlflow==2.9.1
mmcv==1.3.8
mmsegmentation==0.14.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.8.1.78
opencv-python-headless==4.10.0.84
packaging==23.2
pandas==2.1.3
parso==0.8.4
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
plotly==5.22.0
portalocker==2.10.1
prettytable==3.9.0
prompt_toolkit==3.0.47
protobuf==4.25.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==14.0.1
pyarrow-hotfix==0.6
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-hostlist==1.23.0
pytorch-lightning==2.3.3
pytz==2023.3.post1
PyYAML==6.0.1
querystring-parser==1.2.4
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scikit-image==0.24.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.23
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2024.7.21
timm==1.0.7
tokenizers==0.15.0
tomli==2.0.1
# torch==2.1.2+cu121
# torchaudio==2.1.2+cu121
# torchmetrics==1.4.0.post0
# torchvision==0.16.2+cu121
tqdm==4.66.1
traitlets==5.14.3
transformers==4.35.2
triton==2.1.0
ttach==0.0.3
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
websocket-client==1.7.0
Werkzeug==3.0.1
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
